Updated Tuner to properly read and convert float data

+++ app/src/main/java/com/chrynan/android_guitar_tuner/tuner/detection/YINPitchDetector.java
@@ -1,13 +1,13 @@
 package com.chrynan.android_guitar_tuner.tuner.detection;
 
-import com.chrynan.android_guitar_tuner.tuner.AudioConfig;
+import com.chrynan.android_guitar_tuner.tuner.config.AudioConfig;
 
 /**
  * A {@link PitchDetector} implementation that uses a YIN algorithm to determine the frequency of
  * the provided waveform data. The YIN algorithm is similar to the Auto-correlation Function used
  * for pitch detection but adds additional steps to better the accuracy of the results. Each step
  * lowers the error rate further. The following implementation was inspired by
- * <a href="https://github.com/JorenSix/TarsosDSP/blob/master/src/core/be/tarsos/dsp/pitch/Yin.javaz">TarsosDsp</a>
+ * <a href="https://github.com/JorenSix/TarsosDSP/blob/master/src/core/be/tarsos/dsp/pitch/Yin.java">TarsosDsp</a>
  * and
  * <a href="http://recherche.ircam.fr/equipes/pcm/cheveign/ps/2002_JASA_YIN_proof.pdf">this YIN paper</a>.
  * The six steps in the YIN algorithm are (according to the YIN paper):
@@ -28,7 +28,7 @@ public class YINPitchDetector implements PitchDetector {
     // According to the YIN Paper, the threshold should be between 0.10 and 0.15
     private static final float ABSOLUTE_THRESHOLD = 0.125f;
 
-    private final int sampleRate;
+    private final double sampleRate;
     private final short[] resultBuffer;
 
     public YINPitchDetector(final AudioConfig audioConfig) {
@@ -37,7 +37,7 @@ public class YINPitchDetector implements PitchDetector {
     }
 
     @Override
-    public double detect(short[] wave) {
+    public double detect(float[] wave) {
         int tau;
 
         // First, perform the functions to normalize the wave data
@@ -73,7 +73,7 @@ public class YINPitchDetector implements PitchDetector {
      *
      * @param wave The waveform data to perform the AutoCorrelation Difference function on.
      */
-    private void autoCorrelationDifference(final short[] wave) {
+    private void autoCorrelationDifference(final float[] wave) {
         // Note this algorithm is currently slow (O(n^2)). Should look for any possible optimizations.
         int length = resultBuffer.length;
         int i, j;
@@ -81,7 +81,7 @@ public class YINPitchDetector implements PitchDetector {
         for (j = 1; j < length; j++) {
             for (i = 0; i < length; i++) {
                 // d sub t (tau) = (x(i) - x(i - tau))^2, from i = 1 to result buffer size
-                resultBuffer[j] += (short) Math.sqrt(wave[i] - wave[i + j]);
+                resultBuffer[j] += Math.sqrt(wave[i] - wave[i + j]);
             }
         }
     }
